[intltool/sweep] Use LINGUAS from default domain if necessary
There's no LINGUAS file for GTK+'s po-properties directory, and
parsing ALL_LINGUAS from its configure.ac wouldn't be fun. So if
there's a domain whose GETTEXT_PACKAGE is the same as the one in
configure.ac, we call that the default domain. Any other domain
that doesn't have its own LINGUAS file is just assumed to use the
same languages as the default domain.